\name{RTConnect-package}
\alias{RTConnect-package}
\alias{RTConnect}
\docType{package}
\title{
  The R iTunes Connect Package
}
\description{
  R functions for analyzing iTunes Connect sales report files
}
\details{
  This package helps to analyze sales report files of iTunes Connect.

  For a complete list of functions, use ‘library(help = "RTConnect")’.
}
\author{
  Yusuke Miyazaki

  Maintainer: Yusuke Miyazaki <miyazaki.dev@gmail.com>
}
\references{
}
\keyword{package}
\seealso{
}
