\name{RTConnect-package}
\alias{RTConnect-package}
\alias{RTConnect}
\docType{package}
\title{
  The R iTunes Connect Package
}
\description{
  R functions for analyzing iTunes Connect sales report files
}
\details{
  \tabular{ll}{
    Package: \tab RTConnect\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2013-06-10\cr
    License: \tab BSD_3_clause + file LICENSE\cr
  }

  This package helps to analyze sales report files of iTunes Connect.

  For a complete list of functions, use ‘library(help = "RTConnect")’.
}
\author{
  Yusuke Miyazaki

  Maintainer: Yusuke Miyazaki <miyazaki.dev@gmail.com>
}
\references{
}
\keyword{package}
\seealso{
}
